Move files to `internal/components/Errors/`
