+++ packages/next/src/client/components/react-dev-overlay/_experimental/internal/components/Errors/ErrorIndicator/ErrorIndicator.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import { ErrorIndicator } from './ErrorIndicator'
-import { withShadowPortal } from '../../storybook/with-shadow-portal'
+import { withShadowPortal } from '../../../storybook/with-shadow-portal'
 
 const meta: Meta<typeof ErrorIndicator> = {
   title: 'ErrorIndicator',
