+++ packages/next/src/client/components/react-dev-overlay/_experimental/internal/container/BuildError.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import type { VersionInfo } from '../../../../../../server/dev/parse-version-info'
 import { Terminal } from '../components/Terminal'
 import { noop as css } from '../helpers/noop-template'
-import { ErrorOverlayLayout } from '../components/ErrorOverlayLayout/ErrorOverlayLayout'
+import { ErrorOverlayLayout } from '../components/Errors/ErrorOverlayLayout/ErrorOverlayLayout'
 
 export type BuildErrorProps = { message: string; versionInfo?: VersionInfo }
